Fix -Wunused-function in websocket module
